Update to check for Xcode 16 before applying —legacy
